django_shoping/requirements.txt
attrs==17.4.0
coverage==4.4.2
Django==1.11.8
django-rest-framework==0.1.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
pkg-resources==0.0.0
pluggy==0.6.0
py==1.5.2
PyJWT==1.5.3
pytest==3.3.2
pytest-cov==2.5.1
pytest-django==3.1.2
pytz==2017.3
six==1.11.0
